First attempts at flattening/simplifying internals
