+++ lightning_transformers/cli/train.py
@@ -19,9 +19,8 @@ from pytorch_lightning import LightningDataModule
 from pytorch_lightning.utilities.distributed import rank_zero_info
 
 from lightning_transformers.core import TaskTransformer, TransformerDataModule
-from lightning_transformers.core.config import TaskConfig, TrainerConfig, TransformerDataConfig
+from lightning_transformers.core.config import TaskConfig, TokenizerConfig, TrainerConfig, TransformerDataConfig
 from lightning_transformers.core.instantiator import HydraInstantiator, Instantiator
-from lightning_transformers.core.nlp.config import HFTokenizerConfig
 from lightning_transformers.core.utils import set_ignore_warnings
 
 
@@ -32,7 +31,7 @@ def run(
     dataset: TransformerDataConfig = TransformerDataConfig(),
     task: TaskConfig = TaskConfig(),
     trainer: TrainerConfig = TrainerConfig(),
-    tokenizer: Optional[HFTokenizerConfig] = None,
+    tokenizer: Optional[TokenizerConfig] = None,
     logger: Optional[Any] = None,
 ) -> None:
     if ignore_warnings:
