+++ lightning_transformers/core/config.py
@@ -11,13 +11,37 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from dataclasses import dataclass
+from pathlib import Path
+from typing import Optional, Union
+
+from datasets import Version
 
 
 @dataclass
 class TransformerDataConfig:
     batch_size: int = 32
     num_workers: int = 0
+    dataset_name: Optional[str] = None
+    dataset_config_name: Optional[str] = None
+    revision: Optional[Union[str, Version]] = None
+    train_val_split: Optional[int] = None
+    train_file: Optional[str] = None
+    test_file: Optional[str] = None
+    validation_file: Optional[str] = None
+    padding: Union[str, bool] = "max_length"
+    truncation: str = "only_first"
+    max_length: int = 128
+    preprocessing_num_workers: int = 1
+    load_from_cache_file: bool = True
+    cache_dir: Optional[Union[Path, str]] = None
+    limit_train_samples: Optional[int] = None
+    limit_val_samples: Optional[int] = None
+    limit_test_samples: Optional[int] = None
+    train_subset_name: Optional[str] = None
+    validation_subset_name: Optional[str] = None
+    test_subset_name: Optional[str] = None
 
 
 @dataclass
@@ -38,6 +62,20 @@ class TrainerConfig:
 
 
 @dataclass
+class BackboneConfig:
+    pretrained_model_name_or_path: Optional[str] = None
+
+
+@dataclass
 class TaskConfig:
     optimizer: OptimizerConfig = OptimizerConfig()
     scheduler: SchedulerConfig = SchedulerConfig()
+    downstream_model_type: Optional[str] = None
+    backbone: BackboneConfig = BackboneConfig()
+
+
+@dataclass
+class TokenizerConfig:
+    downstream_model_type: Optional[str] = None
+    pretrained_model_name_or_path: Optional[str] = None
+    use_fast: bool = True
