+++ lightning_transformers/core/model.py
@@ -11,42 +11,112 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Dict, Optional, Tuple, Union
+from typing import IO, TYPE_CHECKING, Any, Callable, Dict, Optional, Tuple, Type, Union
 
 import pytorch_lightning as pl
 import torch
 import transformers
+from hydra.utils import get_class
 from pytorch_lightning.utilities import rank_zero_info, rank_zero_warn
+from transformers import PreTrainedTokenizerBase
+from transformers import pipeline as hf_transformers_pipeline
 
-from lightning_transformers.core.config import OptimizerConfig, SchedulerConfig
+from lightning_transformers.core.config import BackboneConfig, OptimizerConfig, SchedulerConfig
 from lightning_transformers.core.instantiator import Instantiator
 
+if TYPE_CHECKING:
+    from transformers import AutoModel, Pipeline
 
-class LitTransformer(pl.LightningModule):
-    """Base class for transformers.
 
-    Provides a few helper functions primarily for optimization.
+class TaskTransformer(pl.LightningModule):
+    """Base class for task specific transformers, wrapping pre-trained language models for downstream tasks. The
+    API is built on top of AutoModel and AutoConfig, provided by HuggingFace.
+
+    see: https://huggingface.co/transformers/model_doc/auto.html
+
+    Args:
+        downstream_model_type: The AutoModel downstream model type.
+            See https://huggingface.co/transformers/model_doc/auto.html
+        backbone: Config containing backbone specific arguments.
+        pretrained_model_name_or_path: Huggingface model to use if backbone config not passed.
+        optimizer: Config containing optimizer specific arguments.
+        scheduler: Config containing scheduler specific arguments.
+        instantiator: Used to instantiate objects (when using Hydra).
+            If Hydra is not being used the instantiator is not required,
+            and functions that use instantiation such as ``configure_optimizers`` has been overridden.
+        tokenizer: The pre-trained tokenizer.
+        pipeline_kwargs: Arguments required for the HuggingFace inference pipeline class.
+        **model_data_kwargs: Arguments passed from the data module to the class.
     """
 
     def __init__(
         self,
-        model: torch.nn.Module,
-        optimizer: Optional[torch.optim.Optimizer] = None,
-        scheduler: Optional[torch.optim.lr_scheduler._LRScheduler] = None,
-    ):
+        downstream_model_type: str,
+        optimizer: OptimizerConfig = OptimizerConfig(),
+        scheduler: SchedulerConfig = SchedulerConfig(),
+        pretrained_model_name_or_path: Optional[str] = None,
+        backbone: Optional[BackboneConfig] = None,
+        instantiator: Optional[Instantiator] = None,
+        tokenizer: Optional[PreTrainedTokenizerBase] = None,
+        pipeline_kwargs: Optional[dict] = None,
+        **model_data_kwargs,
+    ) -> None:
         super().__init__()
-        self.model = model
-        # some optimizers/schedulers need parameters only known dynamically
-        # allow users to override the getter to instantiate them lazily
-        self.optimizer = optimizer
-        self.scheduler = scheduler
+        self.save_hyperparameters()
+        model_cls: Type["AutoModel"] = get_class(downstream_model_type)
+        model_path = backbone.pretrained_model_name_or_path if backbone else pretrained_model_name_or_path
+        self.model = model_cls.from_pretrained(model_path, **model_data_kwargs)
+        self._tokenizer = tokenizer  # necessary for hf_pipeline
+        self._hf_pipeline = None
+        self._hf_pipeline_kwargs = pipeline_kwargs or {}
+        self.instantiator = instantiator
+        self.optimizer_cfg = optimizer
+        self.scheduler_cfg = scheduler
 
     def configure_optimizers(self) -> Dict:
-        """Prepare optimizer and scheduler."""
+        if self.instantiator is None:
+            rank_zero_warn(
+                "You haven't specified an optimizer or lr scheduler. "
+                "Defaulting to AdamW with an lr of 1e-5 and linear warmup for 10% of steps. "
+                "To change this, override ``configure_optimizers`` in the TransformerModule."
+            )
+            optimizer = torch.optim.AdamW(self.parameters(), lr=1e-5)
+            num_training_steps, num_warmup_steps = self.compute_warmup(
+                num_training_steps=-1,
+                num_warmup_steps=0.1,
+            )
+            scheduler = transformers.get_linear_schedule_with_warmup(
+                optimizer, num_warmup_steps=num_warmup_steps, num_training_steps=num_training_steps
+            )
+            return {
+                "optimizer": optimizer,
+                "lr_scheduler": {"scheduler": scheduler, "interval": "step", "frequency": 1},
+            }
+
+        optimizer = self.instantiator.optimizer(self.model, self.optimizer_cfg)
+        # compute_warmup needs the datamodule to be available when `self.num_training_steps`
+        # is called that is why this is done here and not in the __init__
+        self.scheduler_cfg.num_training_steps, self.scheduler_cfg.num_warmup_steps = self.compute_warmup(
+            num_training_steps=self.scheduler_cfg.num_training_steps,
+            num_warmup_steps=self.scheduler_cfg.num_warmup_steps,
+        )
+        rank_zero_info(f"Inferring number of training steps, set to {self.scheduler_cfg.num_training_steps}")
+        rank_zero_info(f"Inferring number of warmup steps from ratio, set to {self.scheduler_cfg.num_warmup_steps}")
+        scheduler = self.instantiator.scheduler(self.scheduler_cfg, self.optimizer)
         return {
-            "optimizer": self.optimizer,
-            "lr_scheduler": {"scheduler": self.scheduler, "interval": "step", "frequency": 1},
+            "optimizer": optimizer,
+            "lr_scheduler": {"scheduler": scheduler, "interval": "step", "frequency": 1},
         }
+
+    def on_save_checkpoint(self, checkpoint: Dict[str, Any]):
+        # Save tokenizer from datamodule for predictions
+        if self.instantiator:
+            checkpoint["instantiator"] = self.instantiator
+
+    def on_load_checkpoint(self, checkpoint: Dict[str, Any]) -> None:
+        if "tokenizer" in checkpoint:
+            self.tokenizer = checkpoint["tokenizer"]
+        self.instantiator = checkpoint.get("instantiator")
 
     @property
     def num_training_steps(self) -> int:
@@ -61,7 +131,7 @@ class LitTransformer(pl.LightningModule):
             num_warmup_steps *= num_training_steps
         return num_training_steps, num_warmup_steps
 
-    def setup(self, stage: str):
+    def setup(self, stage: Optional[str] = None) -> None:
         self.configure_metrics(stage)
 
     def configure_metrics(self, stage: str) -> Optional[Any]:
@@ -71,59 +141,59 @@ class LitTransformer(pl.LightningModule):
         """
         pass
 
+    @property
+    def tokenizer(self) -> Optional["PreTrainedTokenizerBase"]:
+        if (
+            self._tokenizer is None
+            and hasattr(self, "trainer")  # noqa: W503
+            and hasattr(self.trainer, "datamodule")  # noqa: W503
+            and hasattr(self.trainer.datamodule, "tokenizer")  # noqa: W503
+        ):
+            self._tokenizer = self.trainer.datamodule.tokenizer
+        return self._tokenizer
 
-class TaskTransformer(LitTransformer):
-    """Base class for task specific transformers."""
+    @tokenizer.setter
+    def tokenizer(self, tokenizer: "PreTrainedTokenizerBase") -> None:
+        self._tokenizer = tokenizer
 
-    def __init__(
-        self,
-        model: torch.nn.Module,
-        optimizer: Optional[OptimizerConfig] = None,
-        scheduler: Optional[SchedulerConfig] = None,
-        instantiator: Optional[Instantiator] = None,
+    @property
+    def hf_pipeline_task(self) -> Optional[str]:
+        """Override to define what HuggingFace pipeline task to use.
+
+        Returns: Optional string to define what pipeline task to use.
+        """
+        return None
+
+    @property
+    def hf_pipeline(self) -> "Pipeline":
+        if self._hf_pipeline is None:
+            if self.hf_pipeline_task is not None:
+                self._hf_pipeline = hf_transformers_pipeline(
+                    task=self.hf_pipeline_task, model=self.model, tokenizer=self.tokenizer, **self._hf_pipeline_kwargs
+                )
+            else:
+                raise RuntimeError("No task was defined for this model. Try overriding `hf_pipeline_task`")
+        return self._hf_pipeline
+
+    @hf_pipeline.setter
+    def hf_pipeline(self, pipeline: "Pipeline") -> None:
+        self._hf_pipeline = pipeline
+
+    def hf_predict(self, *args, **kwargs) -> Any:
+        return self.hf_pipeline(*args, **kwargs)
+
+    @classmethod
+    def load_from_checkpoint(
+        cls,
+        checkpoint_path: Union[str, IO],
+        map_location: Optional[Union[Dict[str, str], str, torch.device, int, Callable]] = None,
+        hparams_file: Optional[str] = None,
+        strict: bool = True,
+        hf_pipeline_kwargs: Optional[Dict] = None,
+        **kwargs,
     ):
-        super().__init__(model)
-        self.instantiator = instantiator
-        self.optimizer_cfg = optimizer
-        self.scheduler_cfg = scheduler
-
-    def configure_optimizers(self) -> Dict:
-        if self.instantiator is None:
-            rank_zero_warn(
-                "You haven't specified an optimizer or lr scheduler. "
-                "Defaulting to AdamW with an lr of 1e-5 and linear warmup for 10% of steps. "
-                "To change this, either use Hydra configs or override ``configure_optimizers`` in the Task."
-                "For more information: <todo>"
-            )
-            self._set_default_optimizer_scheduler()
-            return super().configure_optimizers()
-
-        self.optimizer = self.instantiator.optimizer(self.model, self.optimizer_cfg)
-        # compute_warmup needs the datamodule to be available when `self.num_training_steps`
-        # is called that is why this is done here and not in the __init__
-        self.scheduler_cfg.num_training_steps, self.scheduler_cfg.num_warmup_steps = self.compute_warmup(
-            num_training_steps=self.scheduler_cfg.num_training_steps,
-            num_warmup_steps=self.scheduler_cfg.num_warmup_steps,
-        )
-        rank_zero_info(f"Inferring number of training steps, set to {self.scheduler_cfg.num_training_steps}")
-        rank_zero_info(f"Inferring number of warmup steps from ratio, set to {self.scheduler_cfg.num_warmup_steps}")
-        self.scheduler = self.instantiator.scheduler(self.scheduler_cfg, self.optimizer)
-        return super().configure_optimizers()
-
-    def _set_default_optimizer_scheduler(self):
-        self.optimizer = torch.optim.AdamW(self.parameters(), lr=1e-5)
-        num_training_steps, num_warmup_steps = self.compute_warmup(
-            num_training_steps=-1,
-            num_warmup_steps=0.1,
-        )
-        self.scheduler = transformers.get_linear_schedule_with_warmup(
-            self.optimizer, num_warmup_steps=num_warmup_steps, num_training_steps=num_training_steps
-        )
-
-    def on_save_checkpoint(self, checkpoint: Dict[str, Any]):
-        # Save tokenizer from datamodule for predictions
-        if self.instantiator:
-            checkpoint["instantiator"] = self.instantiator
-
-    def on_load_checkpoint(self, checkpoint: Dict[str, Any]) -> None:
-        self.instantiator = checkpoint.get("instantiator")
+        model: TaskTransformer = super().load_from_checkpoint(checkpoint_path, map_location, hparams_file, strict)
+        # update model with hf_pipeline_kwargs override
+        if hf_pipeline_kwargs is not None:
+            model._hf_pipeline_kwargs.update(hf_pipeline_kwargs)
+        return model
