--- lightning_transformers/core/nlp/__init__.py
@@ -1,7 +1,0 @@
-from lightning_transformers.core.nlp.config import (  # noqa: F401
-    HFBackboneConfig,
-    HFTokenizerConfig,
-    HFTransformerDataConfig,
-)
-from lightning_transformers.core.nlp.data import HFDataModule  # noqa: F401
-from lightning_transformers.core.nlp.model import HFTransformer  # noqa: F401
