+++ lightning_transformers/task/nlp/multiple_choice/data.py
@@ -15,12 +15,12 @@ from typing import Dict
 
 from transformers import default_data_collator
 
-from lightning_transformers.core.nlp import HFDataModule
+from lightning_transformers.core import TransformerDataModule
 from lightning_transformers.task.nlp.multiple_choice.config import MultipleChoiceDataConfig
 from lightning_transformers.task.nlp.multiple_choice.utils import DataCollatorForMultipleChoice
 
 
-class MultipleChoiceDataModule(HFDataModule):
+class MultipleChoiceDataModule(TransformerDataModule):
     """Defines the ``LightningDataModule`` for Multiple Choice Datasets."""
 
     cfg: MultipleChoiceDataConfig
