+++ lightning_transformers/task/nlp/summarization/data.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from lightning_transformers.core.nlp.seq2seq import Seq2SeqDataModule
+from lightning_transformers.core.seq2seq.data import Seq2SeqDataModule
 from lightning_transformers.task.nlp.summarization.config import SummarizationDataConfig
