+++ lightning_transformers/task/nlp/text_classification/config.py
@@ -1,8 +1,8 @@
 from dataclasses import dataclass
 
-from lightning_transformers.core.nlp import HFTransformerDataConfig
+from lightning_transformers.core.config import TransformerDataConfig
 
 
 @dataclass
-class TextClassificationDataConfig(HFTransformerDataConfig):
+class TextClassificationDataConfig(TransformerDataConfig):
     ...
