--- tests/core/test_model.py
@@ -1,7 +1,0 @@
-import torch
-
-from lightning_transformers.core import LitTransformer
-
-
-def test_lit_transformer():
-    LitTransformer(torch.nn.Linear(1, 1))
